+++ dir-scripts/!init.scm
@@ -1,0 +1,6 @@
+(clear)
+(use-modules (ice-9 readline)) 
+(activate-readline)
+(add-to-load-path ".")
+(use-modules (opencog))
+(use-modules (opencog exec))