+++ dir-scripts/rewrite.scm
@@ -1,0 +1,18 @@
+;Graph rewriting declarations
+(define make-conclusions
+    (Bind
+        ;Declare the variables [optional]
+        (Variable "$object")
+        ;Declare the pattern used to ground the variables
+        (MemberLink
+            (Variable "$object")
+            (Concept "Has mass"))
+        
+        ;If a match is found for the pattern then we want
+        ;to add the following hypergraph at the Atomspace
+        (MemberLink
+            (Variable "$object")
+            (Concept "Has gravity"))))
+
+;Trigger graph rewriting
+(cog-execute! make-conclusions)